pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ambient-tool"
version = "0.1.0"
description = "CLI for pulling and analyzing Ambient Weather data"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "requests",
  "python-dotenv",
  "rich",
]

[project.scripts]
ambient = "ambient_tool.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
